Integrate ethernet vlan code into resurrected network api
